+++ packages/shared/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./scoring.js";
 export * from "./types.js";
+export * from "./unsupportedFormats.js";
 export * from "./urls.js";